Muddled progress?
Things are still breaking in weird ways with the alien updates. I think
the logic just needs to be redesigned from the ground up to better
handle the individual alien movements and increases in speed. The
current design is buggy and leads to errors when aliens are deleted. It
appears that multiple updates are occuring for certain aliens causing them to
toggle into a mismatched animation state. This may occur during
alienTickPeriod updates? Unclear.